Exp2_Bloch_v5_analysis_pipe.py: Changed titles for experiment averages plots to say 'Experiment' (they were previously mislabelled as 'participant'). Exp3_Ricco_v4_analysis_pipe.py: Changed titles for experiment averages plots to say 'Experiment' (they were previously mislabelled as 'participant'). Monitor_response_time_stim.py: Changed fixation time from fps/10 to int(fps/10), not sure it is necessary, but did it anyway.

+++ Nick_scripts/Monitor_response_time_stim.py
@@ -358,7 +358,7 @@ while this_cond < n_conds:
 
     # timing in frames
     # if ISI >= 0:
-    t_fixation =  fps/10  # 1 * fps
+    t_fixation = int(fps/10)  # 1 * fps
     t_interval_1 = t_fixation + probe_duration
     t_ISI = t_interval_1 + ISI
     t_interval_2 = t_ISI + probe_duration
